Move MethodDescriptor out from JavaModuleWrapper
